Fixed bug in time based playlist Fixed bugs in deployment code

+++ sw/libraries/boombox/examples/deployment_timelocked/deployment_timelocked.ino
@@ -165,7 +165,7 @@ void loop()
                 // delay for delayTime milliseconds after trigger has happened. 
                 // This delays playing the sound immediately after trigger 
                 now = millis();
-                while (elapsedTime(now) < (meta.delayTime * 1000))
+                while (elapsedTime(now) < ((uint32_t)meta.delayTime * 1000))
                 {
                     wdt_reset();
                 }
@@ -186,7 +186,7 @@ void loop()
                 // delay for offDelayTime milliseconds. This is the time after playback finishes but we do not allow another sound
                 // to be triggered.
                 now = millis();
-                while (elapsedTime(now) < (meta.offDelayTime * 1000))
+                while (elapsedTime(now) < ((uint32_t)meta.offDelayTime * 1000))
                 {
                     wdt_reset();
                 }
